+++ app/Http/Controllers/Auth/OAuthLoginController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\NotificationMessage\Item;
 use App\NotificationMessage\ItemType;
+use App\Notifications\NewUser;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Auth;
@@ -15,29 +16,31 @@ use Laravel\Socialite\Facades\Socialite;
 
 class OAuthLoginController extends Controller
 {
+    private string $provider = '';
+
     public function google(): RedirectResponse
     {
+        $this->provider = 'google';
+
+        return $this->handleRequest();
+    }
+
+    public function authentik(): RedirectResponse
+    {
+        $this->provider = 'authentik';
+
+        return $this->handleRequest();
+    }
+
+    private function handleRequest(): RedirectResponse
+    {
         try {
-            $user = Socialite::driver('google')->user();
+            $user = Socialite::driver($this->provider)->user();
 
             return $this->tryLogin($user);
         } catch (\InvalidArgumentException $e) {
             Log::error('Exception while logging in.', [$e]);
             NotificationMessage::addErrorNotificationMessage(__('A problem occurred, while logging in.').' '.__('Please try again.'));
-
-            return redirect(route('login'));
-        }
-    }
-
-    public function authentik(): RedirectResponse
-    {
-        try {
-            $user = Socialite::driver('authentik')->user();
-
-            return $this->tryLogin($user);
-        } catch (\InvalidArgumentException $e) {
-            Log::error('Exception while logging in.', [$e]);
-            NotificationMessage::addErrorNotificationMessage(__('A problem occurred, while logging in.'));
 
             return redirect(route('login'));
         }
@@ -74,8 +77,7 @@ class OAuthLoginController extends Controller
 
     private function createUser(string $email, string $name): ?User
     {
-
-        $user = User::create([
+        $user = User::query()->create([
             'name' => $name,
             'email' => $email,
         ]);
@@ -83,6 +85,9 @@ class OAuthLoginController extends Controller
         $user->register();
 
         Log::channel('userManagement')->info('User '.$email.' has been created by Oauth auto registration.');
+        User::getAdmins()->each(function (User $adminUser) use ($user) {
+            $adminUser->notify(new NewUser($user, $this->provider));
+        });
         NotificationMessage::addNotificationMessage(
             new Item(__('User :user created successfully.', ['user' => $name]), ItemType::SUCCESS));